Flatten config-related files into the main kibana pkg
