+++ pkg/license/reporter_test.go
@@ -9,13 +9,6 @@ import (
 	"testing"
 	"time"
 
-	apmv1 "github.com/elastic/cloud-on-k8s/pkg/apis/apm/v1"
-	esv1 "github.com/elastic/cloud-on-k8s/pkg/apis/elasticsearch/v1"
-	kbv1 "github.com/elastic/cloud-on-k8s/pkg/apis/kibana/v1"
-	commonlicense "github.com/elastic/cloud-on-k8s/pkg/controller/common/license"
-	essettings "github.com/elastic/cloud-on-k8s/pkg/controller/elasticsearch/settings"
-	kbconfig "github.com/elastic/cloud-on-k8s/pkg/controller/kibana/config"
-	"github.com/elastic/cloud-on-k8s/pkg/utils/k8s"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
@@ -23,6 +16,14 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	apmv1 "github.com/elastic/cloud-on-k8s/pkg/apis/apm/v1"
+	esv1 "github.com/elastic/cloud-on-k8s/pkg/apis/elasticsearch/v1"
+	kbv1 "github.com/elastic/cloud-on-k8s/pkg/apis/kibana/v1"
+	commonlicense "github.com/elastic/cloud-on-k8s/pkg/controller/common/license"
+	essettings "github.com/elastic/cloud-on-k8s/pkg/controller/elasticsearch/settings"
+	"github.com/elastic/cloud-on-k8s/pkg/controller/kibana"
+	"github.com/elastic/cloud-on-k8s/pkg/utils/k8s"
 )
 
 const operatorNs = "test-system"
@@ -157,7 +158,7 @@ func Test_Get(t *testing.T) {
 						{
 							Name: kbv1.KibanaContainerName,
 							Env: []corev1.EnvVar{{
-								Name: kbconfig.EnvNodeOpts, Value: "--max-old-space-size=2048",
+								Name: kibana.EnvNodeOpts, Value: "--max-old-space-size=2048",
 							}},
 						},
 					},
